requirements.txt
streamlit==1.32.0
langchain==0.1.12
langchain-openai==0.0.8
langchain-community==0.0.27
langgraph==0.0.27
pydantic==2.6.4
python-dotenv==1.0.1
tavily-python==0.3.1
typing-extensions==4.10.0
pytest==8.1.1
black==24.3.0
isort==5.13.2
flake8==7.0.0 